requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
PyPDF2
transformers
accelerate
torch
pydantic
huggingface-hub